[XPipeline]Camera UI - Unify Headers - Output
